backend/requirements.txt
Django
pymysql
djangorestframework
django-cors-headers